refactor: address review feedback, move UniFiAuth to its own module
Moves UniFiAuth (and the exceptions it raises: CannotConnectError,
SslCertificateError, InvalidAuthError) into a new unifi_auth.py, as
requested in review. UniFiClient composes a UniFiAuth instance with no
passthrough properties or delegation shortcuts; internal call sites use
self._auth.authenticated / .method / .invalidate() / .headers() directly.

Production import sites (__init__.py, config_flow.py, coordinator.py) now
import the exceptions from .unifi_auth directly rather than relying on
unifi_client.py's incidental re-export, which also satisfies mypy's
explicit-reexport check.

Test coverage for UniFiAuth (headers, login, api-key verification, method
auto-detection/fallback) moved to a new test_unifi_auth.py, exercised via
make_auth() against mocked HTTP responses instead of monkeypatching
private methods through client._auth. test_unifi_client.py keeps only
tests for UniFiClient's own behaviour (transport, pagination, probe
backoff) plus two tests confirming authenticate() delegates to self._auth.
